requirements.txt
flask==3.0.2
flask-cors==5.0.1
opencv-python==4.9.0.80
mediapipe==0.10.9
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
pyOpenSSL==24.0.0 